DCMAW-8264: Add logging framework and implement for the token lambda (#12)
DCMAW-8264

### What changed
Adds a logger for use in lambdas. This uses AWS Powertools logger:
https://docs.powertools.aws.dev/lambda/typescript/latest/core/logger/

This includes functionality for registering log messages and ensuring
that the lambda can only have access to the registered log messages for
that lambda using the type system.

Currently there are limited properties in the log message - this can be
expanded over time as and when the requirements surface.

This also adds a new ErrorOrValue type, superseeding the LogOrValue
type. The rationale for this is that it feels cleaner to implement logs
at the lambda level and not couple downstream services with the Logging
framework.

For awareness - AWS have native ability now to write structured logs
without the need for a library such as AWS Powertools. This could reduce
the build size and need the dependency ->
https://aws.amazon.com/about-aws/whats-new/2024/07/aws-lambda-search-filter-aggregate-function-logs/.
I'd personally suggest merging as-is and I can work on analysing this
today.

### Why did it change
To improve visibility of incoming requests that do not meet the required
schema/spec and potential defects/server issues within the code.

## Checklists
<!-- Merging this PR is effectively deploying to production. Be mindful
to answer accurately. -->

- [x] There is a ticket raised for this PR that is present in the branch
name
- [x] No PII data logged. [See guidance here]
- [x] Documentation updated