[DCMAW-14492] Include actual and wanted session states in SESSION_CONDITIONAL_CHECK_FAILURE log (#798)
The dynamoDbAdapter is extended to add a sub-object to the data payload of log
message type `LogMessage.UPDATE_SESSION_CONDITIONAL_CHECK_FAILURE`:

```
sessionStates: {
  actual : string,
  valid : Array<string>
}
```

This is used when the DynamoDB adapter catches `ConditionalCheckFailedException`.

The condition check we care about here is that the session is in one of the wanted
states - these vary depending on the type of `UpdateOperation` being executed.

`sessionStates` will contain consistently-formatted SessionState enum strings, so the
two subfields can be compared in Splunk.