DCMAW-16858: fix `backend-api` test failure due to AWS dependency update (#919)
As detailed in #2418, updated versions of certain AWS packages don't play nicely with Jest due to its lack of support for ESM, which means a fix is required before we can update them to enable the tests to run successfully.

Here, we find that specifically the @aws-sdk/client-secrets-manager dependency causes the test:api tests to fail. We therefore apply the fix of setting the --experimental-vm-modules flag when these tests are run, as per #2432.

No other test scripts in either backend-api or test-resources appear to be affected by the issue.