Don't reify GITHUB_TOKEN when generating .npmrc files.
Why: GITHUB_TOKEN is set for us by Github in the environment for a
Github Action.  We use it when checking pull requests, and later
when pushing to main. It's needed in order to pull the private NPM
mobile-id-check-biometric-credential, which is a dependency of
this repo.

When double quote " is used in `echo "$GITHUB_TOKEN"` the value of
the environment variable is substituted (i.e. it is reified) in the
text being echoed.  If this is done when writing to .npmrc we have
the potential to leak an access token.

When single ' is used, the echoed text continues to contain the
`$GITHUB_TOKEN` characters, and no leak occurs.

We must use curly braces, echo '${GITHUB_TOKEN}' in order for NPM
to look to its environment for a token value.

Lastly we must update the workflow jobs to include GITHUB_TOKEN in
the jobs that actually consume the token, and not the job that
constructs the .npmrc file.