[DCMAW-17053] Don't reify GITHUB_TOKEN when generating .npmrc files. (#929)
When double quote " is used in echo "$GITHUB_TOKEN" the value of the environment variable
is substituted (i.e. it is reified) in the text being echoed. If this is done when writing to .npmrc
we have the potential to leak an access token.

When single ' is used, the echoed text continues to contain the $GITHUB_TOKEN characters,
and no leak occurs.

We must use curly braces, echo '${GITHUB_TOKEN}', in order for NPM to look to its
environment for a token value.