Amend `asyncFinishBiometricSessionHandler.ts` to use `persistentIdentifiers` object
A change was introduced in `2.29.0` of the `@aws-lambda-powertools/logger` dependency which breaks how we set persistent identifiers on the logger in `asyncFinishBiometricSessionHandler.ts`; previously, writing an object to the logger if it already had that object on it would merge the two objects, but as of `2.29.0` the new object replaces the old one. This means we can no longer update the persistent identifiers incrementally. We therefore amend `asyncFinishBiometricSessionHandler.ts` to initially create a `persistentIdentifiers` object, which we amend and add to the logger as we go.